feat(account): gate bulk-claim button on multi-stack + not-all-auto-claim
The "Claim funds" shortcut now only enables when the user has payouts
waiting in more than one stack AND auto-claims isn't already on for every
stack (in which case it's redundant). Lift the aggregate auto-claim read
into a shared use-account-auto-claims-state hook so the button and the
toggle read one deduped query.
